Updated by AS 1) removed warnings
git-svn-id: https://sppong.googlecode.com/svn/trunk@280 c580c124-fe00-11dd-b556-9305e745af9e

+++ Source/Base/Includes/Macros.h
@@ -173,7 +173,7 @@ inline int roundUp(int numToRound, int multiple)
 }  
 
 template <class T>
-inline int Sign(const T a)
+inline short Sign(const T a)
 {
 	return a >=0 ? 1 : -1;
 }
